+++ doc/releasenotes.rst
@@ -19,8 +19,8 @@ Release Notes
 -------------------
 
 * Save number of samples taken in calibration step in database.
-* Fix error with reported number of simulation in EpsMixin based samplers.
-* Fix several minor issues.
+* Fix error with reported number of simulations in EpsMixin based samplers.
+* Fix several warnings.
 
 
 0.9.9 (2019-03-25)